% Generated by roxygen2 (4.0.2): do not edit by hand
\name{deepSampleIds2dataFrame}
\alias{deepSampleIds2dataFrame}
\title{deepSampleIds2dataFrame}
\usage{
deepSampleIds2dataFrame(ids)
}
\arguments{
\item{ids}{a vector of sample ids as strings}
}
\value{
A dataframe with columns corresponding to the fields that can be inferred from the names
}
\description{
parses a vector of sample IDs and returns avector with the corresponding fields
}
\examples{
deepSampleIds2dataFrame(c("43_Hm03_BlMa_TO1_WGBS_E","43_Hm05_BlMa_Ct","43_Hm05_BlMa_Ct_NOMe_S","01_HepaRG_LiHR_D32","41_Hf01_LiHe_Ct1_H3K4me1_F"))
}
\author{
Fabian Mueller
}

