% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetadata.R
\name{getDataFrameFromTabFile}
\alias{getDataFrameFromTabFile}
\title{getDataFrameFromTabFile}
\usage{
getDataFrameFromTabFile(fn)
}
\arguments{
\item{fn}{file name}
}
\value{
a dataframe with one row containing relevant sample information
}
\description{
given a two column, tab-separated file, extract relevant information as a data frame
}
\author{
Fabian Mueller
}

