% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sampleIds.R
\name{getDeepSampleIdsFromString}
\alias{getDeepSampleIdsFromString}
\title{getDeepSampleIdsFromString}
\usage{
getDeepSampleIdsFromString(x)
}
\arguments{
\item{x}{string vector in which to look for DEEP Ids}
}
\value{
a list containing the found DEEP IDs for each element in the string vector
}
\description{
looks for occurrences of DEEP IDs in a vector of strings and returns them as a list
}
\examples{
getDeepSampleIdsFromString(c("blubbsad43_Hm03_BlMa_TO1_WGBS_E_1XXX43_Hm05_BlMa_CtNOI43_Hm05_BlMa_Ct","43_Hm05_BlMa_Ct_NOMe_S_2uiadyf01_HepaRG_LiHR_D32","blubb"))
}
\author{
Fabian Mueller
}

