% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetadata.R
\name{getExperimentMdFromFile}
\alias{getExperimentMdFromFile}
\title{getExperimentMdFromFile}
\usage{
getExperimentMdFromFile(fn, assay = NULL, lenient = FALSE)
}
\arguments{
\item{fn}{sample metadata file name}

\item{assay}{assay type. Possible values: 'wgbs','chip','rna','chracc'}

\item{lenient}{do not enforce required fields, just return a warning if one of them is missing}
}
\value{
a dataframe with one row containing relevant sample information
}
\description{
given a experiment metadata file, extract relevant information as a data frame
}
\author{
Fabian Mueller
}

