% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getMetadata.R
\name{getSampleMdFromFile}
\alias{getSampleMdFromFile}
\title{getSampleMdFromFile}
\usage{
getSampleMdFromFile(fn, lenient = FALSE)
}
\arguments{
\item{fn}{sample metadata file name}

\item{lenient}{do not enforce required fields, just return a warning if one of them is missing}
}
\value{
a dataframe with one row containing relevant sample information
}
\description{
given a sample metadata file, extract relevant information as a data frame
}
\author{
Fabian Mueller
}

